Separate lexer and parser
