+++ src/test/java/grammarTest/GrammarTest.java
@@ -1,9 +1,9 @@
 package grammarTest;
 
 import generate.Properties;
-import io.github.therealmone.tdf4j.generator.impl.ParserGenerator;
+import io.github.therealmone.tdf4j.generator.LexerGenerator;
+import io.github.therealmone.tdf4j.generator.ParserGenerator;
 import io.github.therealmone.tdf4j.lexer.Lexer;
-import io.github.therealmone.tdf4j.lexer.LexerFactory;
 import io.github.therealmone.tdf4j.lexer.config.AbstractLexerModule;
 import io.github.therealmone.tdf4j.parser.Parser;
 import io.github.therealmone.tdf4j.parser.config.AbstractParserModule;
@@ -13,7 +13,7 @@ import io.github.therealmone.tdf4j.parser.model.ast.ASTRoot;
 import org.junit.Test;
 
 public class GrammarTest{
-    Lexer lexer = new LexerFactory().withModule(new AbstractLexerModule() {
+    Lexer lexer = LexerGenerator.newInstance().generate(new AbstractLexerModule() {
         public void configure() {
             tokenize("BSL").pattern("^\\[$").priority(1); // int max
             tokenize("BSR").pattern("^\\]$").priority(1); // priority-  необзяательный параметр
@@ -62,7 +62,7 @@ public class GrammarTest{
             tokenize("WS").pattern("\\s|\\n|\\r").priority(Integer.MAX_VALUE).hidden(true);
         }
     });
-    Parser parser = new ParserGenerator().generate(new AbstractParserModule() {
+    Parser parser = ParserGenerator.newInstance().generate(new AbstractParserModule() {
         public void configure() {
             //1
